pyproject.toml
[tool.poetry]
name = "daomatcher-backend"
version = "0.1.0"
description = ""
authors = ["Yab <yabuman46@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
Flask-Cors = "^4.0.1"
gunicorn = "^22.0.0"
gevent = "^24.2.1"
langchain = "^0.2.2"
Flask-SocketIO = "^5.3.6"
Werkzeug = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.7"
bcrypt = "^4.1.3"
python-decouple = "^3.8"
Flask-Mail = "^0.10.0"
PyJWT = "^2.8.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "2.9.9"
pytest = "^7.1.2"
bs4 = "^0.0.2"
gevent-websocket = "^0.10.1"
flask-login = "^0.6.3"
langchain-google-vertexai = "^1.0.5"
langchain-google-genai = "^1.0.3"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]
